aula 3

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma.js";
 
 export async function memoryRoutes(app: FastifyInstance) {
-	app.get("/memories", async () => {
+	app.addHook("preHandler", async (request) => {
+		await request.jwtVerify();
+	});
+
+	app.get("/memories", async (request) => {
 		const memories = await prisma.memory.findMany({
+			where: {
+				userId: request.user.sub,
+			},
 			orderBy: {
 				createdAt: "asc",
 			},
@@ -17,7 +24,7 @@ export async function memoryRoutes(app: FastifyInstance) {
 		}));
 	});
 
-	app.get("/memories/:id", async (request) => {
+	app.get("/memories/:id", async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -29,6 +36,10 @@ export async function memoryRoutes(app: FastifyInstance) {
 				id,
 			},
 		});
+
+		if (!memory.isPublic && memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
 
 		return memory;
 	});
@@ -42,19 +53,17 @@ export async function memoryRoutes(app: FastifyInstance) {
 
 		const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-		const memory = await prisma.memory.create({
+		return prisma.memory.create({
 			data: {
 				coverUrl,
 				content,
 				isPublic,
-				userId: "3d0661c9-0b36-4009-956a-e4a473cdb4e3",
+				userId: request.user.sub,
 			},
 		});
-
-		return memory;
 	});
 
-	app.put("/memories/:id", async (request) => {
+	app.put("/memories/:id", async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -69,7 +78,20 @@ export async function memoryRoutes(app: FastifyInstance) {
 
 		const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-		const memory = await prisma.memory.update({
+		const memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+			select: {
+				userId: true,
+			},
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
+
+		return prisma.memory.update({
 			data: {
 				coverUrl,
 				content,
@@ -79,16 +101,27 @@ export async function memoryRoutes(app: FastifyInstance) {
 				id,
 			},
 		});
-
-		return memory;
 	});
 
-	app.delete("/memories/:id", async (request) => {
+	app.delete("/memories/:id", async (request, reply) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
 
 		const { id } = paramsSchema.parse(request.params);
+
+		const memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+			select: {
+				userId: true,
+			},
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return reply.status(401).send();
+		}
 
 		await prisma.memory.delete({
 			where: {
